+++ src/main/java/api/gbuild/component/button/GButtonWithText.java
@@ -34,8 +34,9 @@ public class GButtonWithText extends GButton {
         
         float[] c = super.rawColor.color();
         textValue.setColor(c[0], c[1], c[2]);
-        textValue.setAlign(PConstants.CENTER);
+        textValue.setAlign(PConstants.LEFT);
         super.content.add(textValue);
+        super.content.dim(super.content.dim().x + 10);
     }
     
     /**
@@ -60,25 +61,51 @@ public class GButtonWithText extends GButton {
         return ((GText)this.content.get(0)).value();
     }
     
+    /**
+     * Specify the mode of the text contained at the button.
+     * Available values are MODEL and SHAPE
+     * 
+     * <p>
+     * Available values are MODEL and SHAPE, so the modification
+     * would not be done if the passed value is not one of these values 
+     * </p>
+     * 
+     * <p>
+     *  For more information about text alignment, click
+     *  <a href="https://processing.org/reference/textMode_.html">here</a>
+     * </p>
+     * 
+     * @param tmode text mode
+     */
+    public void setMode(int tmode) {
+        ((GText)this.content.get(0)).setMode(tmode);
+    }
+    
     @Override
     public void setSize(int size) {
         GText textValue = (GText)(this.content.get(0));
         textValue.setSize(size);
         content.updateComponents();
+        super.content.dim(super.content.dim().x + textValue.pos(true).x);
     }
     
     @Override
     public void draw() {
         if (isVisible()) {
+            this.setSelected(false);
+            
             GText textValue = (GText)(this.content.get(0));
             
             float[] c = super.rawColor.color();
             textValue.setColor(c[0], c[1], c[2]);
+            content.noStroke();
             
             if (manager().mouseX >= pos().x && manager().mouseX <= pos().x + dim().x) {
                 if (manager().mouseY >= pos().y && manager().mouseY <= pos().y + dim().y) {
                     c = super.hoverColor.color();
                     textValue.setColor(c[0], c[1], c[2]);
+                    content.setStrokeColor(c[0], c[1], c[2]);
+                    this.setSelected(true);
                 }
             }