+++ src/main/java/api/gbuild/component/menu/GMenu.java
@@ -1,7 +1,7 @@
 package api.gbuild.component.menu;
 
-import api.gbuild.component.GComponent;
 import api.gbuild.Globals;
+import api.gbuild.component.GComponent;
 import api.gbuild.component.GPanel;
 import api.gbuild.component.button.GButtonOption;
 import processing.core.PApplet;